A few more minor code review changes.

+++ ProductController/CustomProductController.h
@@ -134,7 +134,7 @@ private:
     void InitializeAction( );
     void ProductDependentInitialize();
     void LoadProductConfiguration( ProductPb::ProductConfig& );
-    uint16_t FindThisProductConfig( ProductPb::ProductConfig& );
+    int FindThisProductConfig( ProductPb::ProductConfig& );
     void RegisterLpmEvents();
     void HandleBtLeModuleReady( bool btLeModuleReady );
     void HandleBtLeCapabilityReady( const std::list<std::string>& points );
